breaks down binary search tree into components
