+++ src/components/index.ts
@@ -1,0 +1,8 @@
+export { default as BinarySearchTree } from './BinarySearchTree';
+export { AlgorithmSelector } from './AlgorithmSelector';
+export { TreeVisualizer } from './TreeVisualizer';
+export { ResultsDisplay } from './ResultsDisplay';
+export { AlgorithmExplanation } from './AlgorithmExplanation';
+export { CurrentAlgorithmHeader } from './CurrentAlgorithmHeader';
+export { AlgorithmCard } from './AlgorithmCard';
+export { CodeBlock } from './CodeBlock';
